move get_all_sync_logs_docs to phone.tests.dbaccessors
