+++ corehq/ex-submodules/casexml/apps/phone/tests/utils.py
@@ -1,8 +1,8 @@
 from xml.etree import ElementTree
 from casexml.apps.case.xml import V1
-from casexml.apps.phone.dbaccessors.sync_logs_by_user import get_all_sync_logs_docs
 from casexml.apps.phone.models import get_properly_wrapped_sync_log, get_sync_log_class_by_format
 from casexml.apps.phone.restore import RestoreConfig, RestoreParams, RestoreCacheSettings
+from casexml.apps.phone.tests.dbaccessors import get_all_sync_logs_docs
 from casexml.apps.phone.xml import SYNC_XMLNS
